Move head reference loading and branch deletion to fragment
Head reference will be needed to update branches information displayed inside of fragment.
