+++ app/src/main/res/menu/pull_request_fragment_menu.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <item
+        android:id="@+id/delete_branch"
+        android:title="@string/delete_branch"
+        app:showAsAction="never" />
+
+</menu>
